Disable WASI on CI
This accidentally worked before but `crypto/uid.c` references
nonexistent functions in the WASI SDK which no longer works with more
recent versions of Clang in an updated WASI SDK. Currently WASI SDK I
think has a bug when compiled against the latest version of LLVM (pulled
in with updated rustc) which causes CI failures as well.

I've updated the configuration here to pull in a newer WASI SDK which
exposes the failure to build with `crypto/uid.c`, but I unfortunately
don't have time to track this further right now.