Exclude ci scripts from the published package (#272)
During a dependency review I noticed that the openssl-src package
includes ci scripts in the published crate on crates.io. These scripts
are not required for building the crate and also they are not usfull
outside of the CI setup you use.

This commit excludes CI scripts from the published package.