Add some testing to ci
This commit changes te CI setup to run on all platforms with the
packaged source code instead of the source code in the repository. This
is usful for ensuring that everything required is included. This is the
preparation for excluding more files from the release.

Additionally this bumps the checkout action to version 5.