Remove more files from the published package (#273)
* Add some testing to ci

This commit changes te CI setup to run on all platforms with the
packaged source code instead of the source code in the repository. This
is usful for ensuring that everything required is included. This is the
preparation for excluding more files from the release.

Additionally this bumps the checkout action to version 5.

* Exclude more stuff

This commit excludes more files from the openssl source code from the
code uploaded to crates.io. All of the excluded files are not required
to build openssl on any target listed on ci.

Before: 3699 files, 44.5MiB (10.0MiB compressed)
After: 2383 files, 31.6MiB (6.8MiB compressed)